+++ R/set_error.R
@@ -1,0 +1,6 @@
+# .jrnotes is a hidden environment
+# Setting error to be TRUE allows the check_* function to delay
+# the stop() call to the end of all checks
+set_error = function() {
+  set_error()
+}