+++ NEWS.md
@@ -1,3 +1,6 @@
+# jrNotes (development version)
+  * Internal: Set global error via `set_error()`
+ 
 # jrNotes 0.6.3 _2020-08-27_
   * Feature: check that pkg name and title match
   * Bug: output from check_news now standard